transport: Widen feature negotiation to 128 bits
The required_features path was a single uint64_t end to end, so a
test could not negotiate a feature at bit 64 or above and a
1ULL << 64 shift overflowed. Widen required_features and
virtio_pci_init_features to unsigned __int128 and read and write
all four 32-bit feature words. Bits 0 to 63 negotiate exactly as
before, and bits 64 and above such as RSS_CONTEXT and the tunnel
offloads become expressible.

+++ lib/virtio_pci.h
@@ -118,7 +118,7 @@ int virtio_pci_init(struct virtio_dev *dev);
  * Like virtio_pci_init, but negotiates the requested feature bits
  * (restricted to those the device offers) instead of zero.
  */
-int virtio_pci_init_features(struct virtio_dev *dev, uint64_t wanted);
+int virtio_pci_init_features(struct virtio_dev *dev, unsigned __int128 wanted);
 
 /*
  * Return non-zero if the device offers the given feature bit. Handles
